+++ src/components/helpers/revealPassword.js
@@ -1,0 +1,10 @@
+const revealPass = () => {
+  const pwdField = document.getElementById('password');
+  if (pwdField.type === 'password') {
+    pwdField.type = 'text';
+  } else {
+    pwdField.type = 'password';
+  }
+};
+
+export default revealPass;
